backend/requirements.txt
# Core Django packages
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.7.0

# AI and ML packages
openai==1.75.0
crewai==0.121.1
litellm==1.68.0

# Database and storage
supabase==2.15.2

# PDF processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
asgiref==3.8.1
httpx==0.28.1
typing-extensions==4.13.2

# Build tools
setuptools==78.1.0 